Fix code review issues regarding order shipment handling
ISSUE: PLCORE20-17
