+++ PacklinkPro/Services/BusinessLogic/UserAccountService.php
@@ -1,4 +1,9 @@
 <?php
+/**
+ * @package    Packlink_PacklinkPro
+ * @author     Packlink Shipping S.L.
+ * @copyright  2020 Packlink
+ */
 
 namespace Packlink\PacklinkPro\Services\BusinessLogic;
