+++ PacklinkPro/etc/module.xml
@@ -3,7 +3,7 @@
 /**
  * @package    Packlink_PacklinkPro
  * @author     Packlink Shipping S.L.
- * @copyright  2019 Packlink
+ * @copyright  2020 Packlink
  */
 -->
 <config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
